complated ssr

+++ src/app/com/annaniks/gift4u/views/main/personal-area/personal-area.view.ts
@@ -7,6 +7,7 @@ import { Title } from '@angular/platform-browser';
 import { MatDialog } from '@angular/material';
 import { CookieService } from '../../../services/cookie.service';
 import { TranslateService1 } from '../../../services';
+import { PlatformService } from '../../../services/platform.service';
 
 
 @Component({
@@ -17,14 +18,14 @@ import { TranslateService1 } from '../../../services';
 export class PersonalAreaView implements OnInit {
     private _activeTab: number = 0;
     private _personalAreaItems = [
-        { label: this.translateWord('Personal account','Личный кабинет','Անձնական գրասենյակ'), link: 'user' },
-        { label: this.translateWord('User account','Учетная запись','Հաշիվ'), link: 'account' },
-        { label: this.translateWord('Delivery address','Адреса доставки','Առաքման հասցեները'), link: 'shipping-addresses' },
-        { label:  this.translateWord('My bookmarks','Мои закладки','Նախընտրելիները'), link: 'my-bookmarks' },
-        { label: this.translateWord('Order history','История заказов','Պատվերների պատմություն'), link: 'my-orders' },
-        { label: this.translateWord('Bonus point','Бонусные баллы','Բոնուսային միավորներ'), link: 'bonus-points' },
-        { label: this.translateWord('Payment history','История платежей','Վճարումների պատմություն'), link: 'payment-history' },
-        { label:  this.translateWord('News subscription','Подписка на новости','Նորությունների բաժանորդագրություն'), link: 'newsletter-subscription' },
+        { label: this.translateWord('Personal account', 'Личный кабинет', 'Անձնական գրասենյակ'), link: 'user' },
+        { label: this.translateWord('User account', 'Учетная запись', 'Հաշիվ'), link: 'account' },
+        { label: this.translateWord('Delivery address', 'Адреса доставки', 'Առաքման հասցեները'), link: 'shipping-addresses' },
+        { label: this.translateWord('My bookmarks', 'Мои закладки', 'Նախընտրելիները'), link: 'my-bookmarks' },
+        { label: this.translateWord('Order history', 'История заказов', 'Պատվերների պատմություն'), link: 'my-orders' },
+        { label: this.translateWord('Bonus point', 'Бонусные баллы', 'Բոնուսային միավորներ'), link: 'bonus-points' },
+        { label: this.translateWord('Payment history', 'История платежей', 'Վճարումների պատմություն'), link: 'payment-history' },
+        { label: this.translateWord('News subscription', 'Подписка на новости', 'Նորությունների բաժանորդագրություն'), link: 'newsletter-subscription' },
     ]
 
     constructor(
@@ -35,20 +36,23 @@ export class PersonalAreaView implements OnInit {
         private _mainService: MainService,
         private _title: Title,
         private _matDialog: MatDialog,
-        private _translateService:TranslateService1
+        private _translateService: TranslateService1,
+        private _platformService:PlatformService
     ) {
         this._getUser();
     }
 
     ngOnInit() {
     }
-    public translateWord(key1:string,key2:string,key3:string){
-        return this._translateService.translateImportant(key1,key2,key3)
+    public translateWord(key1: string, key2: string, key3: string) {
+        return this._translateService.translateImportant(key1, key2, key3)
     }
     public onClickLogOut(): void {
-        this._cookieService.remove('accessToken');
-        this._mainService.changeIsAuthorized(false);
-        this._router.navigate(['/']);
+        if (this._platformService.isBrowser) {
+            this._cookieService.remove('accessToken');
+            this._mainService.changeIsAuthorized(false);
+            this._router.navigate(['/']);
+        }
     }
 
     private _getUser(): void {
@@ -83,7 +87,7 @@ export class PersonalAreaView implements OnInit {
     get activateTab(): number {
         return this._activeTab;
     }
-    get language(){
+    get language() {
         return this._translateService.getActiveLanguage()
     }
 }
